Moved assembler to its own module. See notes:
At the time of this commit there is a bug in V0.4 compiler where static methods are not recognized when imported from another module. As a work around, helper functions were added that call the static method inside its own module. This meant I had to change all the calling points.